changing imports of exceptions to function with one class, refactor returning response
