+++ src/chariot/network/Network.py
@@ -1,6 +1,7 @@
 from typing import Dict, List
 from chariot.device.adapter.DeviceAdapter import DeviceAdapter
-from chariot.utility.exceptions import NameNotFoundError, NoIdentifierError, DuplicateNameError, ErrorStrings
+from chariot.utility.exceptions.CustomExceptions import NameNotFoundError, NoIdentifierError, DuplicateNameError
+from chariot.utility.exceptions.ErrorStrings import ErrorStrings
 
 
 class Network:
