+++ src/chariot/utility/PayloadParser.py
@@ -1,4 +1,5 @@
-from chariot.utility.exceptions import NoIdentifierError, ErrorStrings
+from chariot.utility.exceptions.CustomExceptions import NoIdentifierError
+from chariot.utility.exceptions.ErrorStrings import ErrorStrings
 
 
 class PayloadParser:
